+++ src/Application/Bar/DTO/UpdateBarRequest.php
@@ -6,6 +6,9 @@ namespace BarAssistant\Application\Bar\DTO;
 
 final readonly class UpdateBarRequest
 {
+    /**
+     * @param int[] $images
+     */
     public function __construct(
         public int $barId,
         public string $name,
@@ -14,6 +17,9 @@ final readonly class UpdateBarRequest
         public ?string $description = null,
         public ?bool $isPublic = null,
         public ?bool $isInviteCodeEnabled = null,
+        public array $images = [],
+        public ?string $defaultUnits = null,
+        public ?string $defaultCurrency = null,
     ) {
     }
 }